requirements.txt
pandas
anthropic
python-dotenv
flask
flask-cors